# lexical-search/server/server.gen.go
// Package server provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.12.4 DO NOT EDIT.
package server

import (
	"fmt"
	"net/http"
	"time"

	"github.com/deepmap/oapi-codegen/pkg/runtime"
	openapi_types "github.com/deepmap/oapi-codegen/pkg/types"
	"github.com/gin-gonic/gin"
)

// Defines values for MontyDatabaseUsageMode.
const (
	Always    MontyDatabaseUsageMode = "always"
	OnRequest MontyDatabaseUsageMode = "onRequest"
)

// CreateWorkspaceRequest defines model for CreateWorkspaceRequest.
type CreateWorkspaceRequest struct {
	// Name The workspace name
	Name string `json:"name"`

	// WorkTree The absolute workspace path
	WorkTree string `json:"workTree"`
}

// DeleteWorkspaceRequest defines model for DeleteWorkspaceRequest.
type DeleteWorkspaceRequest struct {
	// Id The workspace id
	Id openapi_types.UUID `json:"id"`
}

// Document defines model for Document.
type Document struct {
	// Id A unique identifier for the document
	Id openapi_types.UUID `json:"id"`

	// LineEnd The line number of the last line of text (as a string)
	LineEnd string `json:"lineEnd"`

	// LineStart The line number of the first line of text (as a string)
	LineStart string `json:"lineStart"`

	// Path The path to the parent file relative to the workspace
	Path string `json:"path"`

	// Text The text of the document
	Text string `json:"text"`
}

// Error defines model for Error.
type Error struct {
	Code *string `json:"code,omitempty"`

	// Detail A human-readable explanation specific to this occurrence of the problem.
	Detail string              `json:"detail"`
	Id     *openapi_types.UUID `json:"id,omitempty"`
	Links  *struct {
		About *string `json:"about,omitempty"`
	} `json:"links,omitempty"`
	Status *string `json:"status,omitempty"`
	Title  *string `json:"title,omitempty"`
}

// ErrorsResponse defines model for ErrorsResponse.
type ErrorsResponse struct {
	Errors []Error `json:"errors"`
}

// MontyDatabase defines model for MontyDatabase.
type MontyDatabase struct {
	// Folders A list of folders associated with the Monty database
	Folders []string `json:"folders"`

	// Name The name of the Monty database
	Name      string                 `json:"name"`
	UsageMode MontyDatabaseUsageMode `json:"usageMode"`
}

// MontyDatabaseUsageMode defines model for MontyDatabaseUsageMode.
type MontyDatabaseUsageMode string

// SearchRequest defines model for SearchRequest.
type SearchRequest struct {
	// Databases A list of Monty databases
	Databases []MontyDatabase `json:"databases"`

	// Id The index id
	Id string `json:"id"`

	// Limit The maximum number of results to return
	Limit int `json:"limit"`

	// Query The search query
	Query string `json:"query"`
}

// SearchResponse defines model for SearchResponse.
type SearchResponse struct {
	Results []SearchResult `json:"results"`
}

// SearchResult defines model for SearchResult.
type SearchResult struct {
	Document Document `json:"document"`

	// Score The score of the document
	Score float32 `json:"score"`
}

// Workspace defines model for Workspace.
type Workspace struct {
	// ArchiveDBVersion The version of the archive database
	ArchiveDBVersion int `json:"archiveDBVersion"`

	// BleveDir The abs path of the bleve index
	BleveDir  string    `json:"bleveDir"`
	CreatedAt time.Time `json:"createdAt"`

	// DocCount The number of documents in the bleve index
	DocCount int `json:"docCount"`

	// GitDir The abs path of the .git folder
	GitDir string `json:"gitDir"`

	// Id The workspace id
	Id openapi_types.UUID `json:"id"`

	// LastIndexedCommit The commit hash of the last successfully indexed commit
	LastIndexedCommit string `json:"lastIndexedCommit"`

	// Name The workspace name
	Name      string    `json:"name"`
	UpdatedAt time.Time `json:"updatedAt"`

	// WorkTree The abs path of the workspace folder
	WorkTree string `json:"workTree"`
}

// WorkspacesResponse defines model for WorkspacesResponse.
type WorkspacesResponse struct {
	Workspaces []Workspace `json:"workspaces"`
}

// StreamUpdatesJSONBody defines parameters for StreamUpdates.
type StreamUpdatesJSONBody = []MontyDatabase

// StreamUpdatesParams defines parameters for StreamUpdates.
type StreamUpdatesParams struct {
	// Id The workspace id
	Id openapi_types.UUID `form:"id" json:"id"`
}

// ListWorkspacesParams defines parameters for ListWorkspaces.
type ListWorkspacesParams struct {
	// WorkTree The workspace path
	WorkTree *string `form:"workTree,omitempty" json:"workTree,omitempty"`
}

// SearchJSONRequestBody defines body for Search for application/json ContentType.
type SearchJSONRequestBody = SearchRequest

// StreamUpdatesJSONRequestBody defines body for StreamUpdates for application/json ContentType.
type StreamUpdatesJSONRequestBody = StreamUpdatesJSONBody

// DeleteWorkspaceJSONRequestBody defines body for DeleteWorkspace for application/json ContentType.
type DeleteWorkspaceJSONRequestBody = DeleteWorkspaceRequest

// CreateWorkspaceJSONRequestBody defines body for CreateWorkspace for application/json ContentType.
type CreateWorkspaceJSONRequestBody = CreateWorkspaceRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Search an index
	// (POST /search)
	Search(c *gin.Context)
	// Stream archive updates (SSE) for a workspace
	// (GET /updates)
	StreamUpdates(c *gin.Context, params StreamUpdatesParams)
	// Delete a workspace if it exists
	// (DELETE /workspaces)
	DeleteWorkspace(c *gin.Context)
	// Returns a list of workspaces
	// (GET /workspaces)
	ListWorkspaces(c *gin.Context, params ListWorkspacesParams)
	// Create a new workspace
	// (POST /workspaces)
	CreateWorkspace(c *gin.Context)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// Search operation middleware
func (siw *ServerInterfaceWrapper) Search(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
	}

	siw.Handler.Search(c)
}

// StreamUpdates operation middleware
func (siw *ServerInterfaceWrapper) StreamUpdates(c *gin.Context) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params StreamUpdatesParams

	// ------------- Required query parameter "id" -------------

	if paramValue := c.Query("id"); paramValue != "" {

	} else {
		siw.ErrorHandler(c, fmt.Errorf("Query argument id is required, but not found: %s", err), http.StatusBadRequest)
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "id", c.Request.URL.Query(), &params.Id)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %s", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
	}

	siw.Handler.StreamUpdates(c, params)
}

// DeleteWorkspace operation middleware
func (siw *ServerInterfaceWrapper) DeleteWorkspace(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
	}

	siw.Handler.DeleteWorkspace(c)
}

// ListWorkspaces operation middleware
func (siw *ServerInterfaceWrapper) ListWorkspaces(c *gin.Context) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params ListWorkspacesParams

	// ------------- Optional query parameter "workTree" -------------

	err = runtime.BindQueryParameter("form", true, false, "workTree", c.Request.URL.Query(), &params.WorkTree)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter workTree: %s", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
	}

	siw.Handler.ListWorkspaces(c, params)
}

// CreateWorkspace operation middleware
func (siw *ServerInterfaceWrapper) CreateWorkspace(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
	}

	siw.Handler.CreateWorkspace(c)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router *gin.Engine, si ServerInterface) *gin.Engine {
	return RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router *gin.Engine, si ServerInterface, options GinServerOptions) *gin.Engine {

	errorHandler := options.ErrorHandler

	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.POST(options.BaseURL+"/search", wrapper.Search)

	router.GET(options.BaseURL+"/updates", wrapper.StreamUpdates)

	router.DELETE(options.BaseURL+"/workspaces", wrapper.DeleteWorkspace)

	router.GET(options.BaseURL+"/workspaces", wrapper.ListWorkspaces)

	router.POST(options.BaseURL+"/workspaces", wrapper.CreateWorkspace)

	return router
}